tests/wpt/tests/portals/portals-activate-no-browsing-context.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  assert_implements("HTMLPortalElement" in self);
  let activatePromise = document.createElement('portal').activate();
  await promise_rejects_dom(t, 'InvalidStateError', activatePromise);
}, "A portal with nothing in it cannot be activated");
</script>
